fleshing out pipeline idea for dx api

+++ dx11-tech/inc/Timer.h
@@ -11,7 +11,7 @@ public:
 	};
 
 	Timer();
-	double elapsed(Unit unit = Unit::Seconds) const;
+	float elapsed(Unit unit = Unit::Seconds) const;
 
 private:
 	std::chrono::steady_clock::time_point m_start;
